renaming scal to f144
