+++ src/AccessMessageMetadata/scal/CMakeLists.txt
@@ -1,9 +1,9 @@
-set(scal_SRC
-        scal_Extractor.cpp
+set(f144_SRC
+        f144_Extractor.cpp
         )
 
-set(scal_INC
-        scal_Extractor.h
+set(f144_INC
+        f144_Extractor.h
         )
 
-create_fb_extractor(scal)
+create_fb_extractor(f144)
