+++ src/WriterModule/scal/CMakeLists.txt
@@ -1,9 +1,9 @@
-set(scal_SRC
-  scal_Writer.cpp
+set(f144_SRC
+  f144_Writer.cpp
 )
 
-set(scal_INC
-  scal_Writer.h
+set(f144_INC
+  f144_Writer.h
     )
 
-create_writer_module(scal)
+create_writer_module(f144)
