+++ src/WriterModule/scal/f144_Writer.cpp
@@ -7,7 +7,7 @@
 //
 // Screaming Udder!                              https://esss.se
 
-#include "scal_Writer.h"
+#include "f144_Writer.h"
 #include "MetaData/HDF5DataWriter.h"
 #include "WriterRegistrar.h"
 #include "json.h"
