--- src/tests/WriterModule/scal_WriterTests.cpp
@@ -1,3 +1,0 @@
-//
-// Created by Jonas Nilsson on 2022-06-22.
-//
